Move infrastructure to internal folder
