+++ internal/infrastructure/grpc/props/fetch-server.proto
